[SPARK-51461] Setup `SparkConnect` Swift package structure and CI to test `build`
### What changes were proposed in this pull request?

This PR aims to setup `SparkConnect` Swift package structure and CI to test `build`.

Note that this is a subset of the initial implementation.
- https://github.com/apache/spark-connect-swift/pull/1

### Why are the changes needed?

To setup the initial package structure with CI build test coverage before adding the actual code. Currently, the following two OSs are tested.
- MacOS 15
- Ubuntu 24.04

According to the standard `Swift` package structure,
- https://www.swift.org/getting-started/library-swiftpm/#bootstrapping

this PR adds the following structure for `SparkConnect` package. `SparkConnectError.swift` and `BuilderTests.swift` is added in order to fill the empty directories.
```
$ tree .
.
├── dev
│   └── merge_spark_pr.py
├── LICENSE
├── Package.swift
├── README.md
├── Sources
│   └── SparkConnect
│       └── SparkConnectError.swift
└── Tests
    └── SparkConnectTests
        └── BuilderTests.swift
```

### Does this PR introduce _any_ user-facing change?

No. This is not released yet.

### How was this patch tested?

Pass the CI.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #4 from dongjoon-hyun/SPARK-51461.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>