[SPARK-51472] Add gRPC `SparkConnectClient` actor
### What changes were proposed in this pull request?

This PR aims to add a Swift `SparkConnectClient` actor encapsulating `gRPC` connections which is similar to other language clients.
- Swift (this PR)
```swift
public actor SparkConnectClient {
```

- [Scala](https://github.com/apache/spark/blob/master/sql/connect/common/src/main/scala/org/apache/spark/sql/connect/client/SparkConnectClient.scala):
```scala
private[sql] class SparkConnectClient(
```

- [Python](https://github.com/apache/spark/blob/master/python/pyspark/sql/connect/client/core.py#L597)
```python
class SparkConnectClient(object):
```

This is a part of the following.
- #1

### Why are the changes needed?

To use `gRPC` in the upper `SparkSession` and `DataFrame` layers easily.

### Does this PR introduce _any_ user-facing change?

No, this is not released.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #7 from dongjoon-hyun/SPARK-51472.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>