[SPARK-51736] Make `SparkConnectError` and `StorageLevel` fields public
### What changes were proposed in this pull request?

This PR aims to change the visibility of the following.
- `enum SparkConnectError`
- `StorageLevel` fields

In addition, this PR changes the following test suite to use only `public` APIs of `SparkConnect` by changing the import statement. This will help us validate the visibility change easily.
- `DataFrameTests`
- `DataFrameReaderTests`
- `DataFrameWriterTests`
- `SQLTests`

### Why are the changes needed?

To allow users to use `SparkConnectError` and `StorageLevel`.

### Does this PR introduce _any_ user-facing change?

No. This is a visibility change on the unreleased versions.

### How was this patch tested?

Pass the CIs and manual review.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45 from dongjoon-hyun/SPARK-51736.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>