[SPARK-51815] Add `Row` struct
### What changes were proposed in this pull request?

This PR aims to add `Row` struct and use it.

### Why are the changes needed?

To make `DataFrame` APIs return `Row` instead of `[String?]` in `Swift`.

```swift
- public func collect() async throws -> [[String?]] {
+ public func collect() async throws -> [Row] {

- public func head(_ n: Int32 = 1) async throws -> [[String?]] {
+ public func head(_ n: Int32 = 1) async throws -> [Row] {
```

Note that `Row` is added to support general type fields, but this PR replaces the existing API's `[String?]` signature into `Row`-based signature. The detailed one-to-one mapping among other types will be handled later.

### Does this PR introduce _any_ user-facing change?

Yes, but this is a change to the unreleased version.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #63 from dongjoon-hyun/SPARK-51815.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>