[SPARK-51851] Refactor to use `withGPRC` wrappers
### What changes were proposed in this pull request?

This PR aims to refactor to use `withGRPC` wrappers in `SparkConnectClient` and `DataFrame`.

### Why are the changes needed?

This is helpful not only for reducing the duplicated code but also introducing a security feature.

### Does this PR introduce _any_ user-facing change?

No, this is an internal code refactoring.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

This patch had conflicts when merged, resolved by
Committer: Dongjoon Hyun <dongjoon@apache.org>

Closes #73 from dongjoon-hyun/SPARK-51851.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>