[SPARK-51967] Use `discardableResult` to prevent unnecessary warnings
### What changes were proposed in this pull request?

This PR aims to use `discardableResult` to prevent unnecessary warnings.

### Why are the changes needed?

To simplify the `Spark Connect Swift` API usages.

For example, `count()` API is used to execute an INSERT query. By adding `discardableResult`, we can make it clear that the result is discadable.

```swift
discardableResult
public func count() async throws -> Int64 {
```

### Does this PR introduce _any_ user-facing change?

This is a change to the unreleased version.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #95 from dongjoon-hyun/SPARK-51967.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>