[SPARK-52340] Update `ArrowWriter(Helper)?` and `ProtoUtil` with GH-43170
### What changes were proposed in this pull request?

This PR aims to update `ArrowWriter`, `ArrowWriterHelper`, and `ProtoUtil` to apply the upstream Apache Arrow change, GH-43170.

### Why are the changes needed?

We need to keep syncing with the upstream in order to use Apache Arrow Swift library when it's released. This is a part of preparation to remove these files from this repository in the end.

### Does this PR introduce _any_ user-facing change?

No behavior change.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #182 from dongjoon-hyun/SPARK-52340.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>