[SPARK-52678] Update `ArrowReader.swift` with GH-54
### What changes were proposed in this pull request?

This PR aims to update `ArrowReader.swift` with `GH-54`.

Please note that we cannot use `Apache Arrow Swift` directly yet because
- It is still unable to support `Decimal`.
  - https://github.com/apache/arrow/pull/46628
- It still uses `GRPC Swift` while we use `GRPC Swift 2`.

### Why are the changes needed?

To be ready for the upstream release because `v21.0.0-rc0` tag is created.
- https://github.com/apache/arrow-swift/releases/tag/v21.0.0-rc0

According to the release notes, this seems to be the only one which needs to catch up.

### Does this PR introduce _any_ user-facing change?

No behavior change because this is a refactoring to suppress warning messages.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #207 from dongjoon-hyun/SPARK-52678.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>