[SPARK-53571] Add `integration-test-mac-spark4-iceberg` GitHub Action job
### What changes were proposed in this pull request?

This PR aims to add `integration-test-mac-spark4-iceberg` GitHub Action job. In addition, the existing GitHub Action job `integration-test-mac-iceberg` is renamed to `integration-test-mac-spark3-iceberg` to be clear.

### Why are the changes needed?

Apache Iceberg 1.10 is released with Apache Spark 4 support. We need to have a test coverage for this combination.

https://github.com/apache/iceberg/releases/tag/apache-iceberg-1.10.0

### Does this PR introduce _any_ user-facing change?

No, this is a test infra change.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #239 from dongjoon-hyun/SPARK-53571.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>